requirements.txt
asgiref==3.6.0
attrs==22.2.0
coverage==7.1.0
Django==4.1.6
django-braces==1.15.0
django-mock-queries==2.1.7
djangorestframework==3.14.0
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==17.0.0
image==1.5.33
inflection==0.5.1
iniconfig==2.0.0
mock==5.0.1
model-bakery==1.10.1
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
pytest-mock==3.10.0
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
